requirements-dev.txt
-r requirements.txt # warto dodać linię -r requirements.txt,
                    # aby instalując wersję deweloperską,
                    # automatycznie pobrać też bazowe paczki.
pylint==4.0.4
pylint-django==2.7.0
pylint-plugin-utils==0.9.0
astroid==4.0.3
isort==7.0.0
mccabe==0.7.0
colorama==0.4.6
dill==0.4.1
platformdirs==4.5.1
tomlkit==0.14.0
